Avoid SyntaxError from the browser.
index.html contains a few html templates wrapped in <script> tags.
The browser tries to parse these as javascript, finds HTML and issues
a SyntaxError.  Adding a type="" argument the browser [1] doesn't know
how to parse will make the browser ignore the tag, but still keeping
it available for javascript to read.

In theory HTML5 has a <template> element for this purpose, but I saw
other errors when using that.

[1] Tested with Firefox 24 on Ubuntu.